+++ lib/screens/settings.dart
@@ -332,7 +332,7 @@ Widget buildSettingsDrawer(
                       mainAxisSize: MainAxisSize.min,
                       children: [
                         Text(
-                          "Release 1.1.1",
+                          "Release 1.1.2",
                           style: GoogleFonts.manrope(
                             fontSize: Responsive.font(context, 11),
                             color: Colors.white.withAlpha(80),